requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
Django==3.1
djangorestframework==3.11.1
filelock==3.0.12
importlib-metadata==1.7.0
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.30
zipp==3.1.0
